Change file structure & update types
