--- integrations/eleventy.cjs
@@ -1,1 +1,0 @@
-module.exports = require("./eleventy.mjs").default;
