+++ integrations/eleventy/index.cjs
@@ -1,0 +1,1 @@
+module.exports = require("./index.mjs").default;
