+++ integrations/liquid/index.mjs
@@ -1,7 +1,7 @@
 import { evalToken, Liquid, Tokenizer, toPromise } from "liquidjs";
-import { apiLoadedPromise, CloudCannon } from "../../helpers/cloudcannon.mjs";
-import { registerEleventyBuiltins } from "./11ty-builtins.mjs";
+import { enhanceLiquidError } from "./errors.mjs";
 import { inMemoryFs } from "./fs.mjs";
+import { collectionsProxy, pageProxy } from "./globals.mjs";
 import { group, groupEnd, log } from "./logger.mjs";
 import { createPairedShortcodeTag, createShortcodeTag } from "./shortcodes.mjs";
 
@@ -12,10 +12,14 @@ export { group, groupEnd, log, setVerbose } from "./logger.mjs";
 let sharedLiquidEngine = null;
 
 /**
- * Creates and configures the shared Liquid engine instance.
+ * Creates and configures the shared Liquid engine instance. Host-agnostic
+ * — the host (Eleventy, Jekyll, …) is expected to wire up its own filters,
+ * shortcodes, and built-in ports after this call via the registration
+ * functions exported below. The returned engine should be passed to any
+ * host-side wiring (e.g. `registerEleventyBuiltins(engine)`).
  *
- * @param {{componentDirs?: string[]}} options - Liquid engine options
- * @returns {void}
+ * @param {import("liquidjs").LiquidOptions} [options] - Spread into `new Liquid(...)` (`root`, `extname`, `strictFilters`, etc.)
+ * @returns {import("liquidjs").Liquid}
  */
 export function createSharedLiquidEngine(options) {
   log("Creating shared Liquid engine");
@@ -31,9 +35,6 @@ export function createSharedLiquidEngine(options) {
   });
   log("Liquid engine instantiated");
 
-  registerEleventyBuiltins(sharedLiquidEngine);
-  log("Eleventy built-ins registered (filters + RenderPlugin shims)");
-
   log(
     "Available files in window.cc_files:",
     Object.keys(window.cc_files || {}),
@@ -44,6 +45,8 @@ export function createSharedLiquidEngine(options) {
     createIncludeWithTag(sharedLiquidEngine),
   );
   log("includeWith tag registered");
+
+  return sharedLiquidEngine;
 }
 
 /**
@@ -373,175 +376,3 @@ function createComponentRenderer(name, templateSource) {
   };
 }
 
-/**
- * Inspects a LiquidJS error and returns a new Error with a more
- * descriptive, actionable message for the editable-region error card.
- *
- * @param {unknown} err - The original error
- * @param {string} componentName - The component or template being rendered
- * @returns {Error} An enhanced error with a user-friendly message
- */
-function enhanceLiquidError(err, componentName) {
-  const message = err instanceof Error ? err.message : String(err);
-
-  const unknownFilter = message.match(/undefined filter[:.]?\s*(\S+)/i);
-  if (unknownFilter) {
-    const filterName = unknownFilter[1];
-    return new Error(
-      `Unknown filter "${filterName}" while rendering "${componentName}". ` +
-        `Please check your config and make sure you have registered "${filterName}" in the filters option.`,
-    );
-  }
-
-  const missingTemplate = message.match(/ENOENT.*?"([^"]+)"/);
-  if (missingTemplate) {
-    const filePath = missingTemplate[1];
-    return new Error(
-      `Failed to find included template "${filePath}" while rendering "${componentName}". ` +
-        `Please check that the file exists and is within your configured component directories.`,
-    );
-  }
-
-  const missingTag = message.match(/tag "?(\S+?)"? not found/i);
-  if (missingTag) {
-    const tagName = missingTag[1];
-    return new Error(
-      `Unknown tag "${tagName}" while rendering "${componentName}". ` +
-        `Please check your config and make sure you have registered "${tagName}" in the tags, shortcodes, or pairedShortcodes option.`,
-    );
-  }
-
-  return new Error(`Error rendering "${componentName}": ${message}`);
-}
-
-// Proxy exposed as the `collections` global on the Liquid engine. Property
-// reads kick off an async fetch via the CloudCannon live-editing API; Liquid
-// awaits the returned thenable wherever the value is consumed (for-loops,
-// filter args, output expressions).
-const collectionsProxy = new Proxy(
-  {},
-  {
-    get(_target, key) {
-      if (typeof key !== "string") return undefined;
-      return resolveCollection(key);
-    },
-  },
-);
-
-/**
- * Lazily resolves a CloudCannon collection by key into an array of items
- * suitable for Liquid templates. Each item is realized to plain data so
- * Liquid filters/iteration work without further awaiting nested fields.
- *
- * Item shape is best-effort 11ty-ish (`url`, `inputPath`, `data`); the CC
- * collection layout is not guaranteed to match what Eleventy would have
- * produced.
- *
- * @param {string} key
- * @returns {Promise<Array<{url: any, inputPath: string, data: any}>>}
- */
-async function resolveCollection(key) {
-  await apiLoadedPromise;
-  const collection = CloudCannon.collection(key);
-  const files = await collection.items();
-  return Promise.all(
-    files.map(async (file) => {
-      const data = await file.data.get();
-      return {
-        url: file.url ?? file.path,
-        inputPath: file.path,
-        data,
-      };
-    }),
-  );
-}
-
-// Proxy exposed as the `page` global on the Liquid engine. Mirrors 11ty's
-// `page` object as closely as we can from the Visual Editor API; properties we
-// can't derive from `currentFile()` resolve to `undefined`. Like
-// `collectionsProxy`, property reads return Promises that liquidjs awaits.
-const pageProxy = new Proxy(
-  {},
-  {
-    get(_target, key) {
-      // Guard against accidental thenable detection: if liquidjs (or any
-      // awaiting code) inspects `page.then`, returning a Promise here would
-      // make the proxy itself look like a thenable and trigger recursive
-      // resolution.
-      if (key === "then") return undefined;
-      if (typeof key !== "string") return undefined;
-      return resolvePageProperty(key);
-    },
-  },
-);
-
-/**
- * Resolves a single property of the `page` global against the currently
- * edited file in the Visual Editor. Returns undefined for properties we
- * can't reliably reconstruct without Eleventy build-time state
- * (`outputPath`, `templateSyntax`, `lang`).
- *
- * @param {string} key
- * @returns {Promise<any>}
- */
-async function resolvePageProperty(key) {
-  await apiLoadedPromise;
-  const file = CloudCannon?.currentFile?.();
-  if (!file) return undefined;
-
-  const inputPath = file.path;
-
-  switch (key) {
-    case "inputPath":
-      return inputPath;
-    case "fileSlug": {
-      const base = inputPath.split("/").pop() ?? "";
-      return stripExtension(base);
-    }
-    case "filePathStem": {
-      const stem = stripExtension(inputPath).replace(/^\.?\//, "/");
-      return stem.startsWith("/") ? stem : `/${stem}`;
-    }
-    case "outputFileExtension":
-      return "html";
-    case "date": {
-      const data = (await file.data.get()) ?? {};
-      const raw = /** @type {any} */ (data).date;
-      if (!raw) return undefined;
-      const d = new Date(raw);
-      return Number.isNaN(d.getTime()) ? undefined : d;
-    }
-    case "url": {
-      const data = (await file.data.get()) ?? {};
-      const permalink = /** @type {any} */ (data).permalink;
-      if (typeof permalink === "string") return permalink;
-      return deriveDefaultUrl(inputPath);
-    }
-    default:
-      return undefined;
-  }
-}
-
-/**
- * 11ty's default folder-style permalink: every input path becomes a URL with a
- * trailing slash, with `index` files mapping to the parent directory. Used as
- * the fallback when no front matter `permalink` is set.
- *
- * @param {string} inputPath
- * @returns {string}
- */
-function deriveDefaultUrl(inputPath) {
-  const stem = stripExtension(inputPath).replace(/^\.?\//, "/");
-  const withLeadingSlash = stem.startsWith("/") ? stem : `/${stem}`;
-  const withoutIndex = withLeadingSlash.replace(/\/index$/, "/");
-  return withoutIndex.endsWith("/") ? withoutIndex : `${withoutIndex}/`;
-}
-
-/**
- * Strips the file extension from a path.
- * @param {string} p
- * @returns {string}
- */
-function stripExtension(p) {
-  return p.replace(/\.[^./]+$/, "");
-}
